moves order calculation into its own class
